[Refactor] Move feature extractor methods to utils
