+++ resspect/utils/bump_utils.py
@@ -1,6 +1,6 @@
 """
 # Author: Etienne Russeil and Emille E. O. Ishida
-#    
+#
 # created on 2 July 2022
 #
 # Licensed GNU General Public License v3.0;
@@ -22,6 +22,7 @@ from iminuit.cost import LeastSquares
 
 __all__ = ['protected_exponent', 'protected_sig', 'bump', 'fit_bump']
 
+
 def protected_exponent(x):
     """
     Exponential function : cannot exceed e**10
@@ -34,38 +35,36 @@ def protected_sig(x):
     """
     Sigmoid function using the protected exponential function
     """
-    return 1/(1+protected_exponent(-x))
+    return 1 / (1 + protected_exponent(-x))
 
 
 def bump(x, p1, p2, p3):
     """ Parametric function, fit transient behavior
         Need to fit normalised light curves (divided by maximum flux)
-    
+
     Parameters
     ----------
-    x : np.array 
+    x : np.array
         Array of mjd translated to 0
     p1,p2,p3 : floats
         Parameters of the function
-        
+
     Returns
     -------
     np.array
         Fitted flux array
     """
-    
+
     # The function is by construction meant to fit light curve centered on 40
     x = x + 40
-    
-    return protected_sig(p1*x + p2 - protected_exponent(p3*x))
 
+    return protected_sig(p1 * x + p2 - protected_exponent(p3 * x))
 
 
 def fit_bump(time, flux, fluxerr):
-    
     """
     Find best-fit parameters using iminuit least squares.
-    
+
     Parameters
     ----------
     time : array_like
@@ -76,26 +75,26 @@ def fit_bump(time, flux, fluxerr):
         error in response variable (flux)
     Returns
     -------
-    output : np.ndarray of floats 
+    output : np.ndarray of floats
         Array is [p1, p2, p3, time_shift, max_flux]
-        
+
          p1, p2, p3 are best fit parameter values
          time_shift is time at maximum flux
          max_flux is the maximum flux
-         
+
     """
-    
+
     # Center the maxflux at 0
-    time_shift = -time[np.argmax(flux)] 
+    time_shift = -time[np.argmax(flux)]
     time = time + time_shift
-    
+
     # The function is by construction meant to fit light curve with flux normalised
     max_flux = np.max(flux)
     flux = flux / max_flux
     fluxerr = fluxerr / max_flux
-    
+
     # Initial guess of the fit
-    parameters_dict = {'p1':0.225, 'p2':-2.5, 'p3':0.038}
+    parameters_dict = {'p1': 0.225, 'p2': -2.5, 'p3': 0.038}
 
     least_squares = LeastSquares(time, flux, fluxerr, bump)
     fit = Minuit(least_squares, **parameters_dict)
@@ -106,10 +105,10 @@ def fit_bump(time, flux, fluxerr):
     parameters = []
     for fit_values in range(len(fit.values)):
         parameters.append(fit.values[fit_values])
-        
+
     parameters.append(time_shift)
     parameters.append(max_flux)
-    
+
     return parameters
