+++ tests/test_bump.py
@@ -16,7 +16,8 @@ def test_bump():
     Test the Bump function evaluation.
     """
 
-
+    from resspect.utils.bump_utils import bump
+    
     time = np.array([0])
     p1 = 0.225
     p2 = -2.5
